==> Code/2.ipynb <==
# Load the specified CSV file
import pandas as pd
df = pd.read_csv(r"C:\Users\Owner\Desktop\Devereux\data\estat_ert_bil_eur_a_en.csv\estat_ert_bil_eur_a_en.csv")
df

# show me all unique entries in column TIME_PERIOD
df['TIME_PERIOD'].unique()

#restrict only my TIME_PERIOD to 2000-2021
df = df[df['TIME_PERIOD'].between(2000, 2021)]
df['TIME_PERIOD'].unique()

#name me all unique entries in currency
df['currency'].unique()

# fitler currency for only these BGN (Bulgaria)

# CZK (Czech Republic)

# DKK (Denmark)

# HUF (Hungary)

# PLN (Poland)

# RON (Romania)

# SEK (Sweden)
df = df[df['currency'].isin(['BGN', 'CZK', 'DKK', 'HUF', 'PLN', 'RON', 'SEK'])]
df
